+++ packages/database/prisma/queries/index.ts
@@ -1,4 +1,5 @@
 export * from "./notifications";
+export * from "./operators";
 export * from "./organizations";
 export * from "./purchases";
 export * from "./users";